refactor: extract Hedera client setup logic into reusable utility
- Created `getClientForTests` in `client-setup.ts` to centralize and validate Hedera client initialization.
- Updated HBAR transfer integration tests to use the new utility for cleaner and more maintainable setup.
